+++ R/plot.R
@@ -6,10 +6,12 @@ NULL
 #' @method plot PCOA
 plot.PCOA <- function(x, ..., axes = c(1, 2), labels = FALSE,
                       extra_quali = NULL, extra_quanti = NULL,
+                      ellipse = NULL, hull = NULL,
                       color = NULL, fill = FALSE, symbol = FALSE, size = c(1, 6),
                       xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                       ann = graphics::par("ann"), frame.plot = TRUE,
-                      panel.first = NULL, panel.last = NULL) {
+                      panel.first = NULL, panel.last = NULL,
+                      legend = list(x = "topleft")) {
   ## Prepare data
   arkhe::assert_type(axes, "numeric")
   arkhe::assert_length(axes, 2)
@@ -35,10 +37,7 @@ plot.PCOA <- function(x, ..., axes = c(1, 2), labels = FALSE,
     if (!isFALSE(size)) cex <- khroma::palette_size_sequential(range = size)(extra_quanti)
   }
   ## Highlight qualitative information
-  if (is.null(extra_quali) && length(x@groups) > 0) {
-    extra_quali <- x@groups
-  }
-  if (!isFALSE(extra_quali) && length(extra_quali) > 0) {
+  if (length(extra_quali) > 0) {
     arkhe::assert_length(extra_quali, n)
     if (!isFALSE(color)) col <- khroma::palette_color_discrete(colors = color)(extra_quali)
     if (!isFALSE(fill)) bg <- khroma::palette_color_discrete(colors = fill)(extra_quali)
@@ -82,6 +81,22 @@ plot.PCOA <- function(x, ..., axes = c(1, 2), labels = FALSE,
     )
   }
 
+  if (length(extra_quali) > 0) {
+    ## Add ellipse
+    if (is.list(ellipse) && length(ellipse) > 0) {
+      args_ell <- list(x = x, group = extra_quali, axes = axes,
+                       color = color, fill = FALSE, symbol = FALSE)
+      ellipse <- modifyList(args_ell, val = ellipse)
+      do.call(viz_ellipses, ellipse)
+    }
+    ## Add convex hull
+    if (isTRUE(hull)) {
+      args_hull <- list(x = x, group = extra_quali, axes = axes,
+                        color = color, fill = FALSE, symbol = FALSE)
+      do.call(viz_hull, args_hull)
+    }
+  }
+
   ## Evaluate post-plot and pre-axis expressions
   panel.last
 
@@ -104,6 +119,14 @@ plot.PCOA <- function(x, ..., axes = c(1, 2), labels = FALSE,
       ylab = colnames(coord)[axes[[2]]]
     )
   }
+
+  ## Legend
+  coord <- data.frame(
+    extra_quanti = if (length(extra_quanti) > 0) extra_quanti else rep(NA, n),
+    extra_quali = if (length(extra_quali) > 0) extra_quali else rep(NA, n),
+    cex = cex, col = col, bg = bg, pch = pch, lty = rep(NA, n)
+  )
+  prepare_legend(coord, legend, points = TRUE, lines = FALSE)
 
   invisible(x)
 }