+++ inst/tinytest/test_pca.R
@@ -9,6 +9,7 @@ expect_error(pca(row_zeros))
 
 res <- pca(countries, center = TRUE, scale = FALSE, rank = 5)
 expect_stdout(show(res), "Principal Components Analysis")
+expect_equal(dim(res), 5L)
 expect_equal(rownames(res), rownames(countries))
 expect_equal(colnames(res), colnames(countries))
 expect_equal(dimnames(res), dimnames(countries))
@@ -37,6 +38,9 @@ expect_equal_to_reference(get_inertia(res, margin = 2), file = "_snaps/pca_col_i
 
 # Eigenvalues
 expect_equal_to_reference(get_eigenvalues(res), file = "_snaps/pca_eigenvalues.rds")
+
+# Correlations
+expect_equal_to_reference(get_correlations(res), file = "_snaps/pca_correlations.rds")
 
 # Predict new coordinates ======================================================
 cts <- matrix(data = sample(1:10, 100, TRUE), ncol = 5)